% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{jacSim}
\alias{jacSim}
\title{Calculate jaccard similarity}
\usage{
jacSim(sig, sigL)
}
\arguments{
\item{sig}{Input signature.}

\item{sigL}{List of BugSigDB signatures}
}
\value{
A data.frame
}
\description{
Calculate jaccard similarity
}
