% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_similarity_function.R
\name{simFun}
\alias{simFun}
\title{Calculate jaccard similarity}
\usage{
simFun(sig, sigL)
}
\arguments{
\item{sig}{Input signature.}

\item{sigL}{List of BugSigDB signatures}
}
\value{
A data.frame
}
\description{
Calculate jaccard similarity
}
